Refactor into more idomatic rust
