+++ src/lib.rs
@@ -6,28 +6,48 @@ mod tests {
     use searching;
 
     #[test]
-    fn binary_search() {
+    fn linear() {
+        use searching;
+        let index = searching::linear_search(&"a", &vec!["a", "b", "c", "d", "google", "zoo"]);
+        assert_eq!(index, Some(0));
+
+        let mut index = searching::linear_search(&4, &vec![1, 2, 3, 4]);
+        assert_eq!(index, Some(3));
+
+        index = searching::linear_search(&3, &vec![1, 2, 3, 4]);
+        assert_eq!(index, Some(2));
+
+        index = searching::linear_search(&2, &vec![1, 2, 3, 4]);
+        assert_eq!(index, Some(1));
+
+        index = searching::linear_search(&1, &vec![1, 2, 3, 4]);
+        assert_eq!(index, Some(0));
+
+        index = searching::linear_search(&5, &vec![1, 2, 3, 4]);
+        assert_eq!(index, None);
+    }
+    #[test]
+    fn binary() {
+        use searching;
+        let index = searching::binary_search(&"a", &vec!["a", "b", "c", "d", "google", "zoo"]);
+        assert_eq!(index, Some(0));
+
         let mut index = searching::binary_search(&4, &vec![1, 2, 3, 4]);
-        assert_eq!(index, 3);
+        assert_eq!(index, Some(3));
 
         index = searching::binary_search(&3, &vec![1, 2, 3, 4]);
-        assert_eq!(index, 2);
+        assert_eq!(index, Some(2));
 
         index = searching::binary_search(&2, &vec![1, 2, 3, 4]);
-        assert_eq!(index, 1);
+        assert_eq!(index, Some(1));
 
         index = searching::binary_search(&1, &vec![1, 2, 3, 4]);
-        assert_eq!(index, 0);
+        assert_eq!(index, Some(0));
 
         index = searching::binary_search(&5, &vec![1, 2, 3, 4]);
-        assert_eq!(index, -1);
-
-        index = searching::binary_search(&"a", &vec!["a", "b", "c", "d", "google", "zoo"]);
-        assert_eq!(index, 0);
-
-        index = searching::binary_search(&"google", &vec!["a", "b", "c", "d", "google", "zoo"]);
-        assert_eq!(index, 4);
+        assert_eq!(index, None);
     }
+
     use sorting;
     #[test]
     fn quick_sort() {
